Moved :.:../bin:../../bin:../../../bin:../../../../bin:${HOME}/bin to the end of PATH. Othervise one can stand in ones root directiory running different standard programs like for example 'rm', which in that particular case will be found in ../../../bin i.e. /bin. First of all this is the wrong 'rm', the standard one is the one in /usr/bin, secondly this information is cashed, so when one wants to run 'rm' in another directory it can no longer be found in ../../../bin, and one have a problem.
I also introduced an if statement in order not to get
${PATH}:.:../bin:../../bin:../../../bin:../../../../bin:${HOME}/bin
twice in the PATH if one runs xterm from another shell.

Why do we want to add ${HOME}/bin to the PATH? If the user have any
programs in ${HOME}/bin he will allready have it in his path and the
sourceme file will only add it an extra time. If, on the other hand,
the user don't have any programe in his ${HOME}/bin, it don't have to be in
PATH.

Why do we want .: in PATH? It's a security risk, and it should be up to
the user whether he wants it or not.

+++ sourceme.sh
@@ -4,13 +4,16 @@
 #  . $HOME/f90/pencil_modular/sourceme.sh
 #  into your .bashrc
 #
-CDPATH="./:../:../../:../../../:$HOME"
-PATH=".:../bin:../../bin:../../../bin:../../../../bin:${HOME}/bin:${PATH}"
-IDL_PATH="+../idl:+../../idl:+../../../idl:tmp:+$HOME/idl:<IDL_DEFAULT>"
+if [ -z  $sourceme ]; then 
+  CDPATH="./:../:../../:../../../:$HOME"
+  PATH="${PATH}:.:../bin:../../bin:../../../bin:../../../../bin:${HOME}/bin"
+  IDL_PATH="+../idl:+../../idl:+../../../idl:tmp:+$HOME/idl:<IDL_DEFAULT>"
 
-export CDPATH PATH IDL_PATH
+  export CDPATH PATH IDL_PATH
 
-# alias local='cp -p $1 tmp.$$; \rm $1; mv tmp.$$ $1; chmod u+w $1'
-local() {
+  # alias local='cp -p $1 tmp.$$; \rm $1; mv tmp.$$ $1; chmod u+w $1'
+  local() {
   cp -p $1 tmp.$$; \rm $1; mv tmp.$$ $1; chmod u+w $1;
-}
+  }
+  export sourceme="set"
+fi
